feat(01-05b): execute enrichment pipeline and create final dataset
- Add main_enrich() function to scripts/06_enrich_external.py
- Load cleaned dataset from data/interim/02_cleaned.parquet
- Handle external data with null model names (proceeds with base dataset)
- Add derived analysis columns (5 new metrics)
- Calculate enrichment coverage statistics (96.81% for intelligence-based metrics)
- Save final enriched dataset to data/processed/ai_models_enriched.parquet (188 rows, 16 columns)
- Add --enrich flag to script for enrichment-only execution
- Model tier classification: unknown (127), high (24), mini (20), low (8), medium (7), xhigh (2)

[Rule 3 - Blocking] Fixed Speed column type casting from String to Float64 for division operation

+++ src/enrich.py
@@ -203,10 +203,10 @@ def add_derived_columns(df: pl.DataFrame) -> pl.DataFrame:
 
     # Speed to intelligence ratio
     if "Speed(median token/s)" in derived_df.columns and "intelligence_index" in derived_df.columns:
-        # Convert Speed column to Float64 for division
+        # Cast Speed column to Float64 first (it may be String type)
         derived_df = derived_df.with_columns([
             pl.when(pl.col("intelligence_index") > 0)
-            .then(pl.col("Speed(median token/s)") / pl.col("intelligence_index"))
+            .then(pl.col("Speed(median token/s)").cast(pl.Float64) / pl.col("intelligence_index"))
             .otherwise(None)
             .alias("speed_intelligence_ratio")
         ])
